Optimize solver reset(), isSat() (#26)
* Optimize solver reset()

For whatever reason, ZSolver::reset() is a very costly operation; this
may be due to some bug in Z3. In the meantime, however, an alternative
implementation using ZSolver::push() and pop() has been provided.
The semantics are unchanged, but execution is considerably faster.

* Optimize isSat()

SMTUtils::isSat() was performing an expensive filter()
operation on every invocation; this has been moved into the location
it's actually needed, getModel().